Use transparent gradient on solid color shuttle
This opens up the possibility of having colors associated with each
position, and having the browser transition smoothly between the colors.
Currently, browsers can't transition a background-image: linear-gradient.
But if the linear-gradient is mostly-transparent white to
mostly-transparent black, with a solid color shining through, the browser
can transition that solid background-color.