Add top margin to the switch
The switch was pushed up against its field title or description with no
breathing room, and it looked bad. Apparently, all field types are
responsible for adding their own top margin (rather than relying on
ExpressionEngine styling to take care of this). It looks like the common
value here is an 11px top margin, so go with that.