Allow keyboard navigation
When the inputs are display:none; browsers do not select them in the tab order.
However, changing to position:absolute; left: -9999px; allows the browser to
select the radio buttons without them being visible.

Additionally, add a little bit of a style for when the control is focused, so
it's obvious to the user. I went with a white dotted box, but a better choice
could certainly be made.

Fixes #5.