Trigger a custom event when Switch is changed
Radio button changes are difficult to observe, because they only trigger
when they become checked. Also, getting the various values of the Switch
is difficult if the user doesn't understand the underlying structure of
the component. For these reasons, add a custom event to easily expose when
the value of the Switch changes, and what its various new options are.

Fixes #3.