Support IE8
IE8 does not support CSS transforms, so, for browsers that don't support
CSS transforms, instead of using the .shuttle/span method and moving that
element, Switch will instead just set the background color on the label
that it is already showing.

This was complicated by the fact that IE8 also doesn't support :checked,
which Switch depends on quite a bit, so there needed to be a hack to allow
us to select the :checked item via CSS in IE8.

Fixes #4.